Cleaning dev

See merge request e20190005676/aglter23!6

+++ src/main/java/um/fds/agl/ter23/services/SubjectService.java
@@ -1,8 +1,5 @@
 package um.fds.agl.ter23.services;
 
-import javax.swing.text.html.Option;
-
-import org.apache.el.stream.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -29,12 +26,17 @@ public class SubjectService {
     subjectRepository.save(subject);
   }
 
-  public void deleteSubject(long id) {
+  public int deleteSubject(long id) {
+    if (!subjectRepository.existsById(id)) {
+      System.out.println("\u001B[31m [log:error]Subject not found for id  \u001B[0m");
+      System.out.println(id);
+      return -1;
+    }
     subjectRepository.deleteById(id);
+    return 0;
   }
 
   public SubjectTER getSubject(long id) {
-    System.out.println("\u001B[31m [log:proof] GET SUBJECT  \u001B[0m");
     if (!subjectRepository.existsById(id)) {
       System.out.println("\u001B[31m [log:error]Subject not found for id  \u001B[0m");
       System.out.println(id);
